requirements.txt
dnspython
pycrypto
